+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 - turboChickens spell now takes effect immediately
 - Added fireImmunitySwap spell
 - Added noPortals spell
+- Added extraFallDamage spell